Move drawBytes and skipBytes to Control.Proxy.ByteString
