+++ Control/Proxy/ByteString.hs
@@ -104,17 +104,24 @@ module Control.Proxy.ByteString (
     fromHandleS,
     toHandleD,
     hGetSomeS,
-    hGetSomeS_
+    hGetSomeS_,
+
+    -- ** @pipes-parse@ functionality
+    drawBytes,
+    passBytes
     ) where
 
+import Control.Monad (when)
 import Control.Monad.Trans.Class (lift)
 import qualified Control.Proxy as P
+import Control.Proxy.Parse (draw, unDraw)
 import Control.Proxy.Trans.State (StateP(StateP))
 import Control.Proxy.Trans.Writer (WriterP, tell)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as BL
 import qualified Data.ByteString.Lazy.Internal as BLI
 import qualified Data.ByteString.Unsafe as BU
+import Data.Foldable (forM_)
 import qualified Data.Monoid as M
 import Data.Int (Int64)
 import Data.Word (Word8)
@@ -737,3 +744,46 @@ hGetSomeS_ h = P.runIdentityK go where
                 bs <- lift $ BS.hGetSome h size
                 size2 <- P.respond bs
                 go size2
+
+-- | @drawBytes n@ returns at-most @n@ bytes from upstream. If upstream cannot
+-- produce @n@ bytes, then everything it could produce is returned.
+drawBytes :: (Monad m, P.Proxy p)
+    => Int
+    -> StateP [Maybe BS.ByteString] p
+        () (Maybe BS.ByteString)
+        b' b
+        m BS.ByteString
+drawBytes = loop id
+  where
+    loop diffBs remainder
+        | remainder <= 0 = return $ BS.concat (diffBs [])
+        | otherwise = do
+            mbs <- draw
+            case mbs of
+                Nothing -> return $ BS.concat (diffBs [])
+                Just bs -> do
+                    let len = BS.length bs
+                    if len <= remainder
+                        then loop (diffBs . (bs:)) (remainder - len)
+                        else do
+                            let (prefix, suffix) = BS.splitAt remainder bs
+                            unDraw suffix
+                            return $ BS.concat (diffBs [prefix])
+
+
+-- @skipBytes n@ skips @n@ bytes from upstream. If upstream does not respond
+-- @n@ bytes, then all responded bytes are skipped and this computation returns.
+passBytes :: (Monad m, P.Proxy p)
+    => Int -> StateP [Maybe BS.ByteString] p () (Maybe BS.ByteString) b' b m ()
+passBytes = loop
+  where
+    loop remainder = when (remainder > 0) $ do
+        mbs <- draw
+        forM_ mbs $ \bs -> do
+            let len = BS.length bs
+            if len <= remainder
+                then loop (remainder - len)
+                else do
+                    let (_, suffix) = BS.splitAt remainder bs
+                    unDraw suffix
+                    return ()
