+++ pipes-bytestring.cabal
@@ -23,8 +23,7 @@ Library
         bytestring >= 0.9.2.1,
         transformers >= 0.2.0.0,
         pipes >= 3.3.0,
-        pipes-handle >= 1.0.0
+        pipes-parse >= 1.0.0
     Exposed-Modules:
         Control.Proxy.ByteString
-        Control.Proxy.Handle.ByteString
     Default-Language: Haskell98
